Allow null value for birth_date so it’s simpler to add values to the “skip” list

+++ app/src/com/ternaryop/phototumblrshare/db/PostTagDAO.java
@@ -48,6 +48,10 @@ public class PostTagDAO extends AbsDAO<PostTag> implements BaseColumns {
 	}
 
 	protected void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
+	    // no need to upgrade
+	    if (newVersion == 2) {
+	        return;
+	    }
         db.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
         onCreate(db);
     }
